requirements.txt
pandas
numpy
python-dotenv
requests
apscheduler
playwright
schedule
backports.zoneinfo; python_version < "3.9"
akshare
yfinance
finvizfinance
